# Conflicts:
#	app/Http/Controllers/NotesController.php
#	app/Http/Controllers/UserController.php
#	app/Listeners/StoreUserLoginHistory.php
#	app/Livewire/Administrators/GroupManagement.php
#	app/Livewire/Administrators/LibrarianManagement.php
#	app/Livewire/Administrators/StudentManagement.php
#	app/Livewire/Administrators/TeacherManagement.php
#	app/Livewire/Administrators/UserLoginLog.php
#	app/Livewire/Notes/ShareNoteRecipientsSelect.php
#	app/Livewire/Students/DocumentGenerator.php
#	app/Livewire/Students/Overview.php
#	app/Livewire/Teachers/StudentLogin.php
#	app/Models/LoginActivity.php
#	app/Models/Teacher.php
#	app/Providers/EventServiceProvider.php
#	app/Services/UserLoginService.php
#	app/Traits/BelongsToSchoolEnhanced.php
#	app/Traits/ShouldScopeSchool.php
#	resources/views/components/layouts/app.blade.php
#	resources/views/notes/show.blade.php
#	routes/administrator.php
#	routes/web.php

+++ database/migrations/2025_06_20_215626_add_role_id_in_users_table.php
@@ -12,12 +12,14 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('role_id')->nullable()->after('role');
-            $table->foreign('role_id')
-                ->references('id')
-                ->on('roles')
-                ->onDelete('set null')
-                ->onUpdate('cascade');
+            if (!Schema::hasColumn('users', 'role_id')) {
+                $table->unsignedBigInteger('role_id')->nullable()->after('role');
+                $table->foreign('role_id')
+                    ->references('id')
+                    ->on('roles')
+                    ->onDelete('set null')
+                    ->onUpdate('cascade');
+            }
         });
     }
